refactor: DeserializeEntryを単独のコマンドとして用いるようにし、命名も分かりやすさのために`write`にした
