+++ UpHateblo.Lib/Commands/WriteEntryCommand.cs
@@ -3,9 +3,10 @@ using UpHateblo.Lib.Entities;
 using YamlDotNet.Serialization;
 using YamlDotNet.Serialization.NamingConventions;
 
-namespace UpHateblo.Lib.Markdown;
+namespace UpHateblo.Lib.Commands;
 
-public static class DeserializeEntry
+// TODO: Source GeneratorベースのYAMLパーサーを導入
+public static class WriteEntryCommand
 {
     private static readonly IDeserializer YamlDeserializer =
         new DeserializerBuilder()
@@ -21,6 +22,15 @@ public static class DeserializeEntry
             (.*)$
             """, RegexOptions.Singleline);
 
+    public static MaybeEntry Run(string content)
+    {
+        var (maybeFrontMatter, body) = content.Separate();
+        var frontMatter = maybeFrontMatter is not null
+            ? maybeFrontMatter.ParseFrontMatter()
+            : new MaybeEntry(null, null, null, null, null, null, null, null, null, null);
+        return frontMatter with { Content = body };
+    }
+
     private static (string? maybeFrontMatter, string body) Separate(this string content)
     {
         var match = FrontMatterRegex.Match(content);
@@ -32,14 +42,5 @@ public static class DeserializeEntry
     private static MaybeEntry ParseFrontMatter(this string frontMatter)
     {
         return YamlDeserializer.Deserialize<MaybeEntry>(frontMatter);
-    }
-
-    public static MaybeEntry Run(string content)
-    {
-        var (maybeFrontMatter, body) = content.Separate();
-        var frontMatter = maybeFrontMatter is not null
-            ? maybeFrontMatter.ParseFrontMatter()
-            : new MaybeEntry(null, null, null, null, null, null, null, null, null, null);
-        return frontMatter with { Content = body };
     }
 }
